#AG-25 refactor to use admin panel

+++ routes/web.php
@@ -15,15 +15,13 @@ Route::get('/', function () {
 })->name('home');
 
 Route::group(['prefix' => 'admin'], function () {
-    Route::prefix('/events')->group(function () {
-        Route::get('/', [EventController::class, 'index'])->name('events.index');
-    });
+    Route::get('/events', [EventController::class, 'index'])->name('admin.events.index');
     Route::prefix('/event')->group(function () {
-        Route::get('/create', [EventController::class, 'create'])->name('event.create');
-        Route::post('/store', [EventController::class, 'store'])->name('event.store');
+        Route::get('/create', [EventController::class, 'create'])->name('admin.event.create');
+        Route::post('/store', [EventController::class, 'store'])->name('admin.event.store');
 
-        Route::get('/{id}', [EventController::class, 'event'])->name('event.show');
-        Route::put('/update/{id}', [EventController::class, 'update'])->name('event.update');
-        Route::delete('/delete/{id}', [EventController::class, 'delete'])->name('event.delete');
+        Route::get('/{id}', [EventController::class, 'show'])->name('admin.event.show');
+        Route::put('/update/{id}', [EventController::class, 'update'])->name('admin.event.update');
+        Route::delete('/delete/{id}', [EventController::class, 'delete'])->name('admin.event.delete');
     });
-})->middleware('checkIfAdmin');
+});
